clear warnings

+++ libfixmath/fix16_sqrt.c
@@ -11,7 +11,7 @@
  */
 fix16_t fix16_sqrt(fix16_t inValue)
 {
-    uint8_t  neg    = (inValue < 0) ? 1U : 0U;
+    uint8_t  neg    = (inValue < 0) ? (uint8_t)1U : (uint8_t)0U;
     uint32_t num    = fix_abs(inValue);
     uint32_t result = 0;
     uint32_t bit;
